+++ test/testProject/index-FactoryFromDirectory.js
@@ -1,5 +1,4 @@
 var SkillVCFactory = require('../../lib/skillVCFactory.js');
-var path = require('path');
 
 exports.handler = function(event, context) {
 	SkillVCFactory.createfromDirectory().handler(event, context);