+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### v0.4.0 - (2016-11-XX)
+
+### New Features
+* Filters, Intent Handlers, and Session Handlers now support async processing within them.  To do async operations in a
+Filter or Session Handler return a ```Promise``` from the method and SkillVC will honor the ```Promise``` until it
+is completed.  For Intent Handlers, no ```Promise``` is require as it just needs to call ```context.succeed()```
+or ```context.fail()``` like any other skill.  SkillVC uses the calls to ```context``` to control an internal ```Promise```
+for the skill.
+
+### Deprecated
+* svContext.filterChainCallback deprecated in favor of a ```Promise```
+
 ### v0.3.3 - (2016-10-30) - BREAKING CHANGES
 
 #### New Features